backend/requirements.txt
python-dotenv
fastapi
uvicorn[standard]
pydantic
aiohttp
websockets
httpx
groq
mediapipe
opencv-python-headless
numpy
Pillow
